+++ src/components/Easyclearinput.vue
@@ -5,9 +5,18 @@
             <slot name="input-before"></slot>
             <span v-if="!hasSlot" style="width: 1%;display: table-cell">&nbsp;</span><!-- 占位元素，用于撑开宽度，原因未知 -->
             <span v-if="icon" class="glyphicon form-control-feedback" :class="iconClass" aria-hidden="true"></span>
-            <span class="clear-it glyphicon glyphicon-remove-circle" :class="{ 'has-icon': icon, 'hide': disabled || readOnly }" aria-hidden="true" @click="handleClear"></span>
+            <span @click="handleClear" class="clear-it glyphicon glyphicon-remove-circle" :class="{ 'has-icon': icon, 'hide': disabled || readOnly }" aria-hidden="true"></span>
             <div class="info-text" :class="infoTextClass">{{ infoText }}</div>
-            <input v-el:input :type="type" class="form-control" :disabled="disabled" :readOnly="readOnly" v-model="value" :placeholder="placeholder" @focus="handleFocus" @blur="handleBlur" @change="onChange" @input="onInput" />
+            <input class="form-control"
+                v-el:input
+                :type="type"
+                :disabled="disabled"
+                :readOnly="readOnly"
+                v-model="value"
+                :placeholder="placeholder"
+                @focus="handleFocus"
+                @blur="handleBlur"
+            />
         </div>
     </div>
 </template>
@@ -85,6 +94,10 @@ const EVENT_DELAY = 128
 
 export default {
     props: {
+        name: {
+            type: String,
+            default: 'vc-easyclearinput' + Date.now()
+        },
         type: {
             type: String,
             default: 'text'
@@ -122,23 +135,15 @@ export default {
             type: String,
             default: ''
         },
-        onInput: {
+        onFocus: { // focus回调
             type: Function,
             default: function () {}
         },
-        onChange: {
+        onBlur: { // blur回调
             type: Function,
             default: function () {}
         },
-        onFocus: {
-            type: Function,
-            default: function () {}
-        },
-        onBlur: {
-            type: Function,
-            default: function () {}
-        },
-        onClear: {
+        onClear: { // onClear回调
             type: Function,
             default: function () {}
         }
@@ -156,10 +161,10 @@ export default {
         if (this.autofocus) {
             this.focusInput()
         }
-        var keys = Object.keys(this._slotContents)
-        this.hasSlot = keys.some((item, index) => {
-            return item === 'input-before'
-        })
+        // 检查是否有用户自定义slot传入(input-before)
+        this.checkSlot()
+        // 检查用户是否内联了不该内联的事件(focus & blur)
+        this.checkEvents()
     },
     computed: {
         optionalWidth () {
@@ -197,6 +202,22 @@ export default {
         }
     },
     methods: {
+        checkSlot () {
+            var keys = Object.keys(this._slotContents)
+            this.hasSlot = keys.some((item, index) => {
+                return item === 'input-before'
+            })
+        },
+        checkEvents () {
+            let focus = this._events.focus
+            if (focus && focus.length > 0) {
+                console.warn('if you want to listen on focus event, please use `:on-focus` callback!')
+            }
+            let blur = this._events.blur
+            if (blur && blur.length > 0) {
+                console.warn('if you want to listen on blur event, please use `:on-blur` callback!')
+            }
+        },
         /**
          * 点击清除按钮
          * 1. blur 2. clear 3.focus again